Replaced some u8 and u16 by u32.

+++ source/objects/actors/Hero/Hero.c
@@ -515,7 +515,7 @@ bool Hero_stopMovingOnAxis(Hero this, int axis)
 }
 
 // check direction
-void Hero_checkDirection(Hero this, u16 pressedKey, char* animation)
+void Hero_checkDirection(Hero this, u32 pressedKey, char* animation)
 {
 	ASSERT(this, "Hero::checkDirection: null this");
