+++ source/screens/TitleScreen/State/TitleScreenState.c
@@ -249,7 +249,7 @@ static bool TitleScreenState_processMessage(TitleScreenState this, void* owner _
 
 		case kKeyPressed:
 		{
-			u16 pressedKey = *((u16*)Telegram_getExtraInfo(telegram));
+			u32 pressedKey = *((u32*)Telegram_getExtraInfo(telegram));
 
 			if(K_STA & pressedKey)
 			{
